fix: properly test webpack 4
note: dev server hmr is not working correctly
