+++ jest.config.js
@@ -8,7 +8,7 @@ module.exports = {
   ]
 }
 
-if (process.env.WEBPACK4) {
+if (process.env.VUE_CLI_USE_WEBPACK4) {
   module.exports.moduleNameMapper = {
     '^webpack$': 'webpack-4',
     '^webpack/(.*)': 'webpack-4/$1'
