+++ packages/@vue/cli-service/__tests__/serve.spec.js
@@ -7,7 +7,7 @@ const create = require('@vue/cli-test-utils/createTestProject')
 const serve = require('@vue/cli-test-utils/serveWithPuppeteer')
 
 const sleep = n => new Promise(resolve => setTimeout(resolve, n))
-test('serve', async () => {
+test.only('serve', async () => {
   const project = await create('e2e-serve', defaultPreset)
 
   await serve(
